refactor:  9 files removed and repeated boilerplate eliminated across 6 scanners
