--- internal/renderers/excel/advisor.go
@@ -1,18 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-package excel
-
-import (
-	"github.com/Azure/azqr/internal/models"
-	"github.com/Azure/azqr/internal/renderers"
-	"github.com/xuri/excelize/v2"
-)
-
-func renderAdvisor(f *excelize.File, data *renderers.ReportData, styles *StyleCache) {
-	renderSheet(f, data, sheetConfig{
-		stageName: models.StageNameAdvisor,
-		sheetName: "Advisor",
-		tableFunc: data.AdvisorTable,
-	}, styles)
-}
